fix service worker unresponsive issue
there is a issue in manifect v3 where service worker becomes inactive
and never becomes active again.
currently we were using bg-script to make the api calls to server
to fetch the dislike details.

In this temporary fix, we move the task of making api calls to
content-script.
even when service worker dies, content-scripts are still injected.